backend/python_model/requirements.txt
flask
flask-cors
ultralytics
Pillow
numpy
gunicorn